+++ awc/CHANGES.md
@@ -1,10 +1,12 @@
 # Changes
 
+
 ## [0.2.7] - 2019-09-25
 
 ### Added
 
-* Remaining getter methods for `ClientRequest`'s private `head` field.
+* Remaining getter methods for `ClientRequest`'s private `head` field #1101
+
 
 ## [0.2.6] - 2019-09-12